Modules added
I separated out the classes into their own module. Fixed a couple minor bugs. Note the enemy MP bar is not updating properly. Needs fix.
